fix: :bug: The input validation on Create New Customer has to be improved, therefore I am deleting, debugging and refactoring some parts of the code.
