+++ .gitignore
@@ -22,3 +22,4 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+src/components/styling/icons/LogoPlaceholder/LogoPlaceholder.module.css.d.ts
