+++ src/components/Card.jsx
@@ -6,12 +6,12 @@ export const Card = ({ children, width, height, paddingBottom }) => {
         flexDirection: "column",
         backgroundColor: "white",
         borderRadius: '8px',
-        height: height,
-        width: width,
+        height,
+        width,
         paddingTop: "30px",
         paddingRight: "30px",
         paddingLeft: "30px",
-        paddingBottom: paddingBottom,
+        paddingBottom,
         justifyContent: "center",
         boxShadow: "0 2px 4px rgba(0, 0, 0, 0.04)",
