--- src/components/styling/icons/LogoPlaceholder.module.css/LogoPlaceholder.module.css.d.ts
@@ -1,16 +1,0 @@
-/* eslint-disable */
-/* prettier-ignore */
-// @ts-nocheck
-/**
- * Generated by vite-css-modules
- * https://npmjs.com/vite-css-modules
- */
-const logoPlaceholder: string;
-
-export {
-	logoPlaceholder
-};
-
-export default {
-	logoPlaceholder
-}
