+++ src/components/styling/icons/LogoPlaceholder/LogoPlaceholder.jsx
