+++ src/components/styling/icons/LogoPlaceholder/LogoPlaceholder.module.css
