--- src/pages/giftcards/Giftcards.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function giftcards() {
-    return (
-        <div>giftcards</div>
-    )
-}
